sambungin database

+++ database/migrations/2025_04_21_140149_remove_salinitas_from_sensors_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('sensors', function (Blueprint $table) {
-            $table->dropColumn('salinitas');
+            if (Schema::hasColumn('sensors', 'salinitas')) {
+                $table->dropColumn('salinitas');
+            }
         });
     }
